refactor(reconciliation): split progress and fence tests
