+++ scripts/ci_backend_test_manifest.json
@@ -541,6 +541,7 @@
         "tests::upstream_reconciliation::reconciliation_fresh_429_mixed_key_allows_healthy_sibling",
         "tests::upstream_reconciliation::reconciliation_no_adjustment_completes_only_the_current_usage_generation",
         "tests::upstream_reconciliation::reconciliation_observation_reports_due_window_without_queue_count",
+        "tests::upstream_reconciliation_observation::reconciliation_observation_reports_due_window_without_queue_count",
         "tests::upstream_reconciliation::reconciliation_rejects_reclaimed_claim_after_remote_fetch",
         "tests::upstream_reconciliation::reconciliation_rejects_reclaimed_scheduled_job_claim",
         "tests::upstream_reconciliation::reconciliation_rejects_usage_generation_changed_during_remote_fetch",
@@ -569,6 +570,7 @@
         "remote_attempt_admission::tests::",
         "tests::upstream_reconciliation_continuation::",
         "tests::upstream_reconciliation_engine::",
+        "tests::upstream_reconciliation_key_observations::",
         "tests::upstream_reconciliation_fencing::",
         "tests::upstream_reconciliation_projection::",
         "tests::upstream_reconciliation::reconciliation_controlled_retry_advances_the_claim_attempt_before_success",
